fixed everything lol...
