--- Features/Visuals.hh
@@ -1,2 +1,0 @@
-#include "Features.hh"
-
