+++ Features/Visuals/Visuals.cc
@@ -1,34 +1,19 @@
-#include "Features.hh"
+#include "Visuals.hh"
 
-#include "../Console/Console.hh"
-#include "../Drawing/Drawing.hh"
-#include "../Entities/CBaseCombatWeapon.hh"
-#include "../Entities/CCSPlayer.hh"
-#include "../Entities/Cacher.hh"
-#include "../Globals/Globals.hh"
-#include "../Memory.hh"
-#include "../SDK/CGlobalVarsBase.hh"
-#include "../SDK/CUserCmd.hh"
-#include "../SDK/Constants.hh"
-#include "../SDK/ILocalize.hh"
-#include "../SDK/IVDebugOverlay.hh"
-#include "../SDK/IVEngineClient.hh"
+#include "../../Console/Console.hh"
+#include "../../Drawing/Drawing.hh"
+#include "../../Entities/CBaseCombatWeapon.hh"
+#include "../../Entities/CCSPlayer.hh"
+#include "../../Entities/Cacher.hh"
+#include "../../Globals/Globals.hh"
+#include "../../Memory.hh"
+#include "../../SDK/CGlobalVarsBase.hh"
+#include "../../SDK/CUserCmd.hh"
+#include "../../SDK/Constants.hh"
+#include "../../SDK/ILocalize.hh"
+#include "../../SDK/IVDebugOverlay.hh"
+#include "../../SDK/IVEngineClient.hh"
 
-static inline bool ValidMoveType() {
-    return !(g_pMemory->LocalPlayer()->m_nMoveType() == MOVETYPE_NOCLIP || g_pMemory->LocalPlayer()->m_nMoveType() == MOVETYPE_LADDER);
-}
-
-void Features::NoDuckDelay(SDK::CUserCmd* pCmd) {
-    if (BOOL_GET(bRef, "misc.no_duck_delay"); bRef)
-        pCmd->m_nButtons |= IN_BULLRUSH;
-}
-
-void Features::BunnyHop(SDK::CUserCmd* pCmd) {
-    if (BOOL_GET(bRef, "misc.bunny_hop"); bRef)
-        if (ValidMoveType())
-            if (pCmd->m_nButtons & IN_JUMP && !(g_pMemory->LocalPlayer()->m_fFlags() & FL_ONGROUND))
-                pCmd->m_nButtons &= ~IN_JUMP;
-}
 
 static bool ComputeBoundingBox(CCSPlayer* pPl, Vector_t<int>::V4& vecOut) {
     Vector_t<float>::V3 vecMins, vecMaxs;
